Implemented PlayerSwapHandItemsListener
Players are able to use the swap hand item feature to break the captcha.

To reproduce (prior to this patch): Join the server, getting a captcha. Press your hotkey to swap items in your hands. Solve the captcha. You will still not be able to move or interact with anything, but the plugin will tell you the captcha has been solved. Dropping the captcha in your hand will give you the same one back.